Zizes again and again

+++ 0x00-hello_world/6-size.c
@@ -1,8 +1,11 @@
 #include <stdio.h>
+
 /**
- * main - Entry point: prints sizes of char, int, long int, etc
+ * main - Entry point prints the size of various types on the
+ * computer it is compiled and run on.
  *
- * return: Always 0 (Success)
+ *
+ * Return: Always 0 (Success)
  */
 
 int main(void)
@@ -13,11 +16,11 @@ int main(void)
 	long long int ln;
 	float f;
 
-	printf("Size of a char: %lu\n", sizeof(c));
-	printf("Size of an int: %lu\n", sizeof(i));
-	printf("Size of a long int: %lu\n", sizeof(l));
-	printf("Size of a long long int: %lu\n", sizeof(ln));
-	printf("Size of a float: %lu\n", sizeof(f));
+	printf("Size of a char: %lu\n",  (unsigned long)sizeof(c));
+	printf("Size of an int: %lu\n", (unsigned long)sizeof(i));
+	printf("Size of a long int: %lu\n", (unsigned long)sizeof(l));
+	printf("Size of a long long int: %lu\n", (unsigned long)sizeof(ln));
+	printf("Size of a float: %lu\n", (unsigned long)sizeof(f));
 
 	return (0);
 }
